code cleaning and clarity [skip ci]
